Edit functions. Splitted into a new file the Pure functions.
